mp3 player/music-visualizer/requirements.txt
eyed3==0.9.7
librosa==0.9.2
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.3.0
pygame==2.1.2
